fix：添加公社订单接口文档,修改部分下单流程

+++ src/main/java/com/moil/hafen/web/domain/LessonOnlineOrder.java
@@ -23,7 +23,7 @@ public class LessonOnlineOrder implements Serializable {
     private Integer lessonId;
     private String lessonName;
     private Double price;
-    private Double hafenCoin;
+    private Integer hafenCoin;
     private Double couponAmount;
     private Integer couponId;
     private String couponName;
