+++ src/main/java/com/moil/hafen/web/service/impl/CommuneTicketOrderServiceImpl.java
@@ -62,7 +62,7 @@ public class CommuneTicketOrderServiceImpl extends ServiceImpl<CommuneTicketOrde
                 actualPayment = actualPayment - coupon.getDiscountPrice();
             }
         }
-        Double hafenCoin = order.getHafenCoin();
+        Integer hafenCoin = order.getHafenCoin();
         if(hafenCoin!=null&&hafenCoin>0){
             actualPayment = actualPayment-hafenCoin;
         }
