+++ src/main/java/com/moil/hafen/web/service/impl/LessonOfflineOrderServiceImpl.java
@@ -79,7 +79,7 @@ public class LessonOfflineOrderServiceImpl extends ServiceImpl<LessonOfflineOrde
                 actualPayment = actualPayment - coupon.getDiscountPrice();
             }
         }
-        Double hafenCoin = order.getHafenCoin();
+        Integer hafenCoin = order.getHafenCoin();
         if(hafenCoin!=null&&hafenCoin>0){
             actualPayment = actualPayment-hafenCoin;
         }
